ws2_32: getaddrinfo("", ...) on OS X.
getaddrinfo() on Windows resolves an empty, but non-null host string to the addresses of all the network interfaces. A workaround for Linux was added in 75be2284, but unfortunately it does not work on OS X, since there, the host part of the machine name alone usually fails to resolve.

Tested on OS X 10.7.2, Arch Linux (Kernel 3.1.4, glibc 2.14.1) and Ubuntu Oneric.

See http://klickverbot.at/p/getaddrinfo for a detailed comparison of the edge case behavior on the different OSes.