ws2_32: getaddrinfo(NULL, "", ...) on OS X.
If getaddrinfo() is passed a null host and an empty port string, it still resolves to 127.0.0.1/0.0.0.0 on Windows and Linux, but fails on OS X. The added workaround (rewriting the port to "0") works on both Linux and OS X.

Tested on OS X 10.7.2, Arch Linux (Kernel 3.1.4, glibc 2.14.1) and Ubuntu Oneric, and verified against Windows Server 2008 R2.

See http://klickverbot.at/p/getaddrinfo for a detailed comparison of the edge case behavior on the different OSes.